Update patch set 3
Patch Set 3: Code-Review+1

This gets the job done, but a bettter solution might be to use a text view instead of a list view and copy the contents of the im model to the text view. Since the order and content of sent messages should never change, this seems like a nice solution and the user can then copy any part of the text and multiple messages at the same time.

This is also how the KDE, gnome, and mac clients ended up implementing this.

Patch-set: 3