Update patch set 3
Patch Set 3:

I don't know about Mac or Gnome but in KDE he did pretty much the same thing as i did except that he is inheriting QListView to setup the model and signals directly in constructor. As for enabling multi mesage copy it would be trivial with the current implementation. The only limitations is the copy part of the text thing but should we really go trough the hassle of somewhat breaking the model view architectures only for that ?

Patch-set: 3
Label: Code-Review=0