Update patch set 5
Patch Set 5: Code-Review-2

Please don't add logic like without consultation with the rest of the team.

This has implications for the user and should be homogeneous across clients. Neither the OSX nore the Gnome clients do that. 

And we want to show that the call failed or had an error, because we want the user to know what's happening.

This is still a potential activable option (off by default), but it will need some discussion.

Patch-set: 5
Reviewer: Gerrit User 8 <8@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-2