Update patch set 2
Patch Set 2:

You should take a higher dpi version. Also maybe we should switch to png because QtSvg is deprecated and the scaling isn't good because of the way they do it. They load the image at the original size and then scale the bitmap... hence the horrible effect when we upscale

Patch-set: 2
Label: Verified=0