ui: display error on creating / adding exising account
when creating or adding existing ring account,
in the case of wrong PIN or password, the UI spinner cycled infinitely.

No cycling anymore; Added ERROR case handling in switch of endSetup().
(beware of the sketchy #undef because of prior #define ERROR, which i didn't found.)

Added an error page on wizardDialog stacked widget, which closes the wizard on Ok clicked().

Change-Id: I762e2ecbe36b2962c15f5abe095ffb2ddf4768bb