User is informed that registration can be long
On Windows, daemon can take some time to create an account and it
is even worse when a user want to register its name. A label is
now displayed to advise the user about that. Moreover, this label is
modified during the process to show which step is currently
running.

Change-Id: Ia32099f9f46f430293be9e0c03828ea12728e5ce