display the selected account's RingID/username
This patch improves the way the client react to a modification
of the selected account.

From a UI point of view, the displayed Ring ID now depends
on the selected account in the combo box. Previously, it was
always set to the first enabled Ring account found in
AccountModel.

The old behavior is removed from findRingAccount() which now
only focus on looking if an account needs migration or if it
is needed to create one (at first launch).

Change-Id: If63a74eb320b8fc31064d87b04b5e062fc0cc10a