add delegate for pending contact requests
This change adds an item delegate used to display every
elements in the pending contact requests list.

The displayed view is consistent with the existing smart
list view. It shows the RingID of the sender and its picture.
This last one is currently the default picture but is expected
to become the one embedded in the payload if any.

Change-Id: Idc7afb5c6bfcda7bce67dfd042a28f9789506815