setup: detect a previous install of Ring
When Ring is installed, some keys are set in the Registry to indicate
various locations, including the uninstall.exe and the root install
folder.

Now, the installer check the presence of this last value and use it as
the default install location. With this change, updating Ring is just
about pressing “Next” even if it is installed in a custom location.

Change-Id: I48b28c840c094c94f03e763c72f29d833aaa690b
Tuleap: #1460