use new lrc system for account selection + ui fixes
Recent changes in LRC introduced a selection system to choose a
current active account in AccountModel. It was decided to prefer the
selectionModel of the AvailableAccountModel as it filters out accounts
not enabled.

Moreover, the selected account is now the one used when looking for a
new contact. The default scheme used depends on the account type and
the account linked to the contact method is now the selected one.

Finally, the CurrentAccountWidget now better reflects changes of
profile's picture and selected account when this one is changed by an
external event (like a call on an enabled, but not selected, account).

Change-Id: I5ad50a6f6b9d28e03ab031ffecfb0018965475f9