fix ContactMethod/Account association
Adding a new contact by its registered username was correctly linking
it to the current selected account but adding it with a raw RingID
wasn't doing this step. A consequence of this was a fail when sending
a contact request to a new contact added by RingID.

The account associated to a ContactMethod is now set in both cases.

Moreover, a method was added to retrieve the current selected account
as it is often needed in CallWidget.

Change-Id: Ie1476ad85a5e624d7684a95fad37abc01a2d23ee