display vCard data in contact request view
The whole contact request view design has been modified to show every
informations contained in vCard that are useful for a user to decide
if he should or not trust somebody. It is also better looking than
the previous version.

Those informations are presented:
 * An embedded picture if any (with a fallback to the default one)
 * The profile name chosen by the sender of the request
 * The "best ID" which is the registered name or the RingID (in that
   order of priority)

This last information is a security measure to check the real identity
of the contact request sender.

Change-Id: Ibcb3e8667f91c8123b9016e6dfd72621bfb4f3f1
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>