fix incorrect behavior of unread message indicator
When the messaging panel was closed by clicking on the back button, a
slot was still triggered when the last selected contact method
received a message. This one was calling setAllRead() from
TextRecording to indicate that all messages were read by the user and
then the indicator wasn't displayed.

Clicking on the back button now disconnect this signal so new incoming
messages are not marked as read.

Change-Id: I65049ac13ea6fe5c758935333ceda0f97f2b45ab
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>