replace text by icons on main tabs + ui fixes
Tabs are now smaller and use icons instead of text for "Conversations"
and "Contact requests" tabs. It gives a better looking UI and room for
 additional pages.

Moreover, tab switching was giving inconsistencies in many situations
between item selected on the left and what is displayed on the right
panel. Showing or hiding a widget on the right panel now depends on
the selection model of the list on the left. Consequently, clicking on
a back button now deselects the item on the left which is what switches
the view back to the welcome page.

Change-Id: Ib67149dba62e15bebcfe3ebce2398b7717e3f85d
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>