change account deletion processus
The account deletion was performed by solely clicking the bin icon
in the configuration account panel of a selected account.
This was done without warning nor confirmation.

This patch moves the dedicated button to the account list bottom
ass in the other desktop clients, and adds a confimation popup,
summarising the account identifiers (in order, alias, registered name,
ringID), and displays a warning about losing the account in case
it hasn't been exported or added to another device.

Change-Id: I259388da9e143400ab65e88d1fb7c223dbfbff19