modernize avatar placeholder
Default avatar displayed for new contacts is replaced by a more modern
color circle with a capital letter inside.

The color palette used is the same as in the GNOME client. It comes
from the material.io website and is a 16 color subset of the full one
proposed by Google.

The letter is the first one of the bestName() corresponding to the
current item. The “Segoe UI” font is used as recommended by Windows
design guidelines.

Change-Id: I4c8c5729a77c0ff52d18de33c8e2fb8b5b1d7a88