implement SmartList model for new LRC api
A SmartListModel class is added as a wrapper of the new LRC model for
conversations. It is meant to replace RecentModel use as the SmartList
model and thus implements a QAbstractItemModel interface.

The main method of this class is data() which returns every pieces of
information we may want to display for any entry of the list (name,
picture, number of unread messages, etc.)

Change-Id: Idff90b6f547f35d9ee06e42a868760c329e6b1d1