define LRC_IMPORT for DLL import on win32
This patch is meant to be used with corresponding LRC patch fixing
DLL symbols export/import. With LRC_IMPORT defined, Q_DECL_IMPORT is
automatically specified on every class of LRC which are, respectively,
declared Q_DECL_EXPORT during LRC build.

One exception is AccountSerializationAdapter which is exported from
the client and not LRC.

This patch fixes new style's connect() (functor-based) which were not
possible at all with signals introduced with LRC. Especially, it
enables connecting signals with lambda's and compile-time checking of
signals connection.

Change-Id: Ie4f68038b2ba1a524c7a63126fcb0660e3d8cd85