Fix first wizard hangup
when closing the first wizard without creating an account
was resulting in a hang up of the application.
this was due to a exit(0) called in the middle of nowhere.
This patch revises a bit the architecture to allow the information to
reach the main and close the app correctly. (passing by return values
because signals do not work without objects).

Change-Id: I8d4b1bcdcf2646342fd7015509ef7773ca4b09d1