check if account is not null in video ui init
For some reason the account linked to a call is sometimes a nullptr.
This situation induces crashes when we try to initialize some part of
the UI according to attributes of the account.

This problem is circumvented by checking if account is null or not and
initializing the ui to a fallback state. A complete fix would require
modifying parts of the LRC which are due to disappear at the end of
the ongoing refactoring process.

Change-Id: Iccedab793f56d69083688b595cd651b18f18f2c7
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>