build: fix build on Fedora >26
Default QMake configuration in mingw packages has changed since
Fedora 26 and some modifications are required in order to make the
ring client able to compile on it:
 - GDI32 lib is now explicitely linked in our project file. It was
   previously done automatically by QMake "win32-g++" spec.
 - libpcre16-0.dll has been replaced by libpcre2-16-0.dll as a Qt
   dependency and must be shipped instead of the former one.
 - Dockerfile is updated for Fedora 27.

With this patch, the client shouldn't be able to compile anymore on
Fedora <26. Moreover, if building with "ring-project" facilities, a
patch is also required in that repository modifying the QMake selected
spec from "win32-g++" to "mingw-w64-g++".

Change-Id: I367312e66d790d82d6aa1aaa1ad3560e1abaca96