message view: "jump to latest" button only clickable on button content.
The "jump to latest" button container was impednig all clicks on
elements behind it (i.e. the ability to click on the three dots
option when you hover over a message). That is now fixed.

Moreover, before this fix you would not be able to scroll up or down
on the chatview if the cursor was in the "jump to latest" container
area. Now that the pointer-events for the container are set to none
this is not an issue anymore.

Gitlab: #484
Change-Id: I009fd1f245474eeddf5c6ddf222c39814dc0a9fd